requirements.txt
feedparser>=6.0.10
openai>=1.40.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
requests>=2.31.0
pytest>=7.0
